Move common behavior to separate package
