+++ gateway-service/src/test/java/org/zowe/apiml/acceptance/CorsPerServiceTest.java
@@ -10,23 +10,12 @@
 package org.zowe.apiml.acceptance;
 
 import io.restassured.http.Header;
-import org.apache.http.ProtocolVersion;
-import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpUriRequest;
-import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.message.BasicHeader;
-import org.apache.http.message.BasicStatusLine;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatchers;
-import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.zowe.apiml.acceptance.netflix.ApimlDiscoveryClientStub;
-import org.zowe.apiml.acceptance.netflix.ApplicationRegistry;
-
-import java.io.IOException;
+import org.zowe.apiml.acceptance.common.AcceptanceTest;
+import org.zowe.apiml.acceptance.common.AcceptanceTestWithTwoServices;
 
 import static io.restassured.RestAssured.given;
 import static org.apache.http.HttpStatus.SC_OK;
