+++ gateway-service/src/test/java/org/zowe/apiml/acceptance/TimeoutPerServiceTest.java
@@ -14,6 +14,8 @@ import org.apache.http.client.methods.HttpRequestBase;
 import org.apache.http.client.methods.HttpUriRequest;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
+import org.zowe.apiml.acceptance.common.AcceptanceTest;
+import org.zowe.apiml.acceptance.common.AcceptanceTestWithTwoServices;
 
 import java.io.IOException;
